Kerem/harden metrics lambda (#8244)
our lambda logs are showing that we're timing out at 10 minutes. other
logs are showing that alchemy is having a hard time keeping up with our
requests. for this reason, i'm increasing the timeout and a Promise.all
between two request batches. each batch is already internally
parallelized and they're sending a huge amount of requests. So this
should relax the lambda and help it to completion