Kerem/notification service subquery fix (#8255)
notification service on omega is giving the attached error. it seems
like the rows from the first query have grown too big to be parameters
for the `notIn` query. that's why we're not observing the error on gamma
or alpha. i don't know a good way to fix this in native prisma so i'm
just yoloing with raw sql, so please take a good look at the query

<img width="1223" alt="image"
src="https://github.com/user-attachments/assets/778df439-c7c8-47c3-9c5e-92455462a82d">