app: OfflineToast (#8347)
This adds a global offline toast using the use-network-status hook.
I know we've had some false/stuck offline states in the past. But hoping
this will show up for some of these bug reports we're getting that
suggest user is offline when trying to perform actions.
Later we can stick this in the WalletReady component if we like to
prevent users from performing actions while offline - noticed Coinbase
mobile app does this.