userops: better userOpsStore, UserOpTxModal: global, read new userOpsStore state (#8360)
This PR does a few things I've been wanting to get to.

More organizing of middleware.

Convert paymasterProxyMiddleware to use new paymaster proxy response
format

Update userOpsStore to hold 
- entire op instead of only gas values
- decoded call data for specific operations
- new error codes returned from paymaster worker requests

The decoded call data allows for extracting relevant data used in the
UserOpTxModal. Meaning we no longer need any of the current props for
that component, and it is now a global component in App.tsx instead of
having to add it for every tx flow! Oh and we also get to inform the
user of why they're paying for the tx.