fix(app): show channel disabled screen (towns-11891) (#8409)
- channel disable screen
- hide disable button, return to channel info pannel to prevent panel
closing abruptly

Note: the original problem we're trying to solve is how to delete
channel. We're changing the problem slightly by "disabling". At this
stage we should ideally make so you the button simply flips to "Enable".
This would require further changes but just saying... I'll leave a note
in linear for Fei too