add script and update ci to always verify upgraded facets on basescan (#8499)
We need to verify on both blockscout and basescan. Our interaction
script that upgrades facets takes in 1 verifier which is set to
blockscout. We need to also verify on basescan as contract names are
read into the diffing pipeline from basescan api only. This PR adds a
script in ci to always verify upgraded facets once upgraded in basescan.