fix Bug Report causing blank screen on public town page (#8544)
https://github.com/HereNotThere/harmony/pull/8518 added some details to
bug report that requires TownsContext, and there are some instances
where the bug report is mounted in a modal outside of TownsContext,
causing an uncaught error and blank screen (outside our top level error
boundary).

So I moved inside the ErrorBoundary, and I just opted to only use the
node status if it's available. I tried putting AppBugReportOverlay
inside TownsContextProvider inside App, and it did seem to work...but I
feel like this must exist outside the context for some reason and I'm
not entirely sure that it's safe to move inside.