fix(app): space to take mute precedence over channel (towns-15208) (#8553)
because we push the spaceId to mutedChannelIds:
https://github.com/HereNotThere/harmony/blob/ac37b2ebc64dcae1831bf020a595db75c32c3150/clients/web/lib/src/client/TownsNotifciationSettings.ts#L504

I'm pushing this because it looks like this is the way we had it?
However I have a doubt this is how it should really behave, I would
probably still like seeing channels bolded if I enter a muted space (but
avoid badges on an app level)? The latter is straight forward to achieve
but don't want to mix things up in this PR.

In the same vain, do we skip space > channel precedence if we edit the
channel setting? I guess not, but should we in this case grey out the
channel settings in muted towns?