crossmint: fix email issue and use a different event for status (#8576)
a few issues:

populates with your email if you are using email or google, otherwise it
will be blank. fixes issue with empty email string causing crossmint
form error

we were using `order?.phase === 'payment'` to track that a user had
submitted the form. But actually this is the default phase when you load
the form, and the next phase to kick in is `delivery`.

I took a look deeper on `order.payment.status` - we can use either
`order.payment.status.complete || order.phase.delivery` to track that
the payment (form) was submitted, and went with the former because it
has another `in-progress` status that maybe they can get working in the
future and we can use that - for me it never was `in-progress`.

Anyway I also prevented the modal from closing once it reaches
`order.payment.status.complete` - there is still a couple seconds
between when you click submit and when the crossmint modal updates to
have their delivery UI in which you can close the modal, but it's a lot
better than allowing the modal to be closed during the duration of the
delivery, and the best we can do for now.