UserOpTxModal: modularize, clean up (#8590)
After Crossmint the UserOpTxModal is bloated, confusing and hard to
reason about. And I think we will introduce further complexity with
tipping and/or adding decent fund wallet ux.
So this PR makes things more composable if we need to show different
kinds of modals. Could go further but I think this is good for now.

Tested:

- [x] join as brand new user
- [x] join as existing user
- [x] join paid town - CC
- [x] join paid town - ETH
- [x] create town
- [x] any other tx