fix: allow free spaces to update other membership settings (#8602)
edit membership form batches a bunch of user operations.
The contract will revert if we pass a new membership price of < the
minimum membership price.

For a space that is created free, the membership price is 0.

Two issues:
- the form would not allow user to submit if the price was 0. Fix was to
allow the form to submit if the fixed price space was originally a price
of 0. Otherwise don't allow
- the batching logic in UserOperations needed to be updated to match
both a fixed price space and price > 0

So as long as a fixed price space with a cost of 0 is submitted with a
new price of 0, the userop logic will exclude updating the price (the
old price is the same as the new price, so no need to update the price),
as the other operations can proceed.

A caveat that has no affordance in the UI is, if you update a free space
to have a price, you can never go back to a free space.