app: tip a message (Base only) (#8610)
This adds an icon to the MessageContextMenu. A dollar icon that when
pressed allows user to tip the message.

All UI is hidden behind a env.VITE_TIPS_ENABLED flag, so this is fine to
merge.

Still awaiting River changes for adding events to the stream. For now,
this only interacts with the contract on Base networks. It will not yet
display anything on the town or message like accumulated tips, etc.

Other work for other PRs:
- mobile
- River events
- accumulated tips on space