fix wrong address mapping in useOfflineStore (#8772)
didn't see this in my original attempt to fix in
https://github.com/HereNotThere/harmony/pull/8752. the bug was that
use-linked-wallets could set the wrong value in the offline store. it
might set linked EOA. so i removed that code

but we also need to invalidate the corrupted cache entry. i tried
several different approaches, this one is the simplest and i think we
can get away with it. essentially the the current user's rootKey ->
abstractAccount mapping will be fetched every app load, a single time
every 24 hours. We can remove this code in a few weeks

I tested by manipulating the persisted mapping, and refreshing the app