userops: handle replacement underpriced (#8813)
## Problem
Replacement underpriced occurs when a user operation becomes "stuck" in
the bundler's mempool, typically because it's gas fees become
underpriced and take longer than expected to mine.

Scenario:
build + send userop to bundler w/ estimated gas fees -> userop in
bundler's mempool -> meanwhile gas fees spike -> userop is now
underpriced -> bundler waits for gas fees to drop low enough for userop
to pass

In order to replace a user operation, a new user op needs to be sent
with an increase of at least 10% in gas fees of the userop sitting in
the mempool.

## Solution
I've found that a replacement underpriced error from Alchemy does not
include the details of the stuck user operation.
Therefore, the userOpStore tracks both current + pending userops.
`current` - userop that is built on client, _prior_ to sending to
bundler
`pending` - a clone of `current`, after it has been sent and accepted in
bundler, containing the userOpHash.

If there is a pending op, during the creation of the user operation gas
fees are estimated to include a > 10% increase of gas fees. If there is
no pending op, the user op just uses the normal gas fee estimations.

This will address the scenarios when a user's tx takes greater than our
wait time, which is currently 30 seconds. The user will get an error,
and once they try the tx again (or another), the replacement logic will
kick in. This does not address automatically retrying the user operation
if it fails to land after the 30 second wait time. That will be added
later.

### Tests
For automated testing, in userops.replacementUnderpriced.test, in order
to create replacement underpriced scenarios, essentially I'm just
sending multiple userops immediately w/o waiting for prior ones to land.
Since block time for the chain is 2 seconds, rapid fire userops will
need to replace previous userops in order to succeed.
Tested a bunch of txs manually in app