app: explore page with background (#8811)
@nachoiacovino saw your
[comment](https://github.com/HereNotThere/harmony/pull/8805#issuecomment-2613442454)
from the explore page.

TLDR; Unless you're doing things on the app layer like today you
shouldn't have to deal with this. But if you have to, the solution is to
use a `<Card />` - if you take a look at the current `<TopBar />` you
will see an example.

The first level for the app is implicit: on desktop we're using `<Card
/>`, and on Mobile mostly `<Panel>` that have a nuance of grey depending
on the context. Skipping an extra explicit level to avoid things like
`<Box background={isMobile ? 'level3' : 'level2' } />` when transferring
a component from a card to a panel etc.

in this PR:
- reusing the same layout as <AppPanelLayout> + <Card> for background
(app store banner too)
- no background for mobile (there's no design for this?)