fix: selective subscribe to useUserLookupStore and stop prefetching in useEnsLookup (#8858)
TownsContextProvider was rendering a ton. The fix was 2 parts:
use-user-lookup-updater had this which was causing over rendering
```
const { ... } = useUserLookup()
```
so updated to
[useShallow](https://github.com/pmndrs/zustand?tab=readme-ov-file#selecting-multiple-state-slices)
```
const { ... } = useUserLookup(useShallow(s => ({})))
```
This was not enough as use-ens-lookup was causing a render loop b/c of
its dependency on lookupUsers.
So I removed the prefetching that the ens lookup was doing - I think
that's fine? We're already running the user lookup which is going to
getEnsData on every user, do we really need to start prefetching all of
these immediately?

I also updated anywhere else using useUserLookup() to useUserLookup(s =>
...)