app: memo context + dlog.info changes (#8855)
this is not a fix but prevents a lot of unnecessary renders. 

while the towns context is already wrapped in a couple of memos, one or
more of the hooks within the implementation is triggering updates
without actually changing the value of the context. The goal would be
the figure out what actually causes this, but this would at the least
prevent the component tree to re-render.

also:

- move all logs to dlogger because it's noisy. Currently too.
- logs changed values (when the sub-tree gets updated)
- remove an odd "hook" logger (?)

demo:

if you run this during join, you will notice a bunch of
`towns:context:info context impl render` logs, those are the ones that
used to trigger an update while "new value" are the ones requiring an
update

<img width="1624" alt="Untitled"
src="https://github.com/user-attachments/assets/76d840d8-c799-49d4-bd9b-2953c605b9cd"
/>