Create useIsHNTMember.ts (#8909)
Warning: Controversial PR 😂 

I would like to get trading running in prod ASAP next week so that we
can all try it out with our prod wallets...
This is a super easy way to simply gate it on who is a member of HNT —
not planning to keep it for long but it'll get the job done.

I want to avoid over-engineering something like this. This took me 5
mins to make and we can move on with trading. Hopefully this piece of
code will stay in the codebase for only a few weeks.

We could of course use launchdarkly or whatever but it will take time to
implement and be much heavier than this.