TOWNS-18379 don’t throttle timeline by default, make it explicit (#8919)
throttling the timeline breaks reactions (when emoji reacting to a
message the reaction will flicker and take a long time)

this splits the api into a “raw” and a “throttled” timeine, with the
intention that devs will choose the throttled version unless they
explicitly need the raw version. I think this hurts perf a little but,
but it still seems fine on my machine.