Validate webhook url and response (#2410)
The app registry server now waits for the user metadata stream to be
established, extracts the first encryption device, and compares that to
the encryption device returned by the webhook when the webhook is called
with an Initialize payload.

In addition, there is a fix here to the http client wrapper that causes
it to work when wrapping an arbitrary transport type. I included that in
this PR because this change removes the need for validation webhook URLs
directly. In addition, this struct contains an upgrade to disallow
non-https traffic and redirects.