Update the streams tracker to allow tracking streams after allocation (#2412)
For some cases, the application utilizing the stream tracker may not be
able to determine if a stream should be tracked upon creation. The
motivating example is the app registry service, which sees an app
registered at some point in time after the app's user inbox stream has
been allocated. In order to avoid storing ciphertexts for every
(session, device key) on the network, let's introduce an alternate path
to manually kick off tracking a stream as soon as the application
determines it needs to be tracked.