Tools for repairing candidate storage on testnet (#2456)
This PR contains two scripts, one written in go and another in python.

The go script will generate a list of streams that have at least one
candidate in storage.

The python script will call GetStream on each stream from the output of
the go tool, forcing the node to load each stream into the cache and
start generating miniblocks for that stream. It will also validate that
the latest miniblock returned by the stream has advanced beyond the
candidates in memory. One the stream advances, the node should clean up
the stream's storage.