app: bump sw update interval 2x (#9110)
there's an issue related to how we poll for service worker updates. I'm
experimenting with a solution here #9109

in the meantime, this PR would serve as a zero-risk proof that the
`main-sw` is the culprit.

by incrementing the poll interval by 10 (from 5 to 50 minutes) we should
in theory reduce the bandwidth noticeably.

I'd like to get this in to next build if possible to see how it plays
out.