Add logs to alpha in order to debug the entitlement failures for delegated wallets (#2589)
We cannot test the delegate wallets feature fully locally because we
rely on delegate contracts deployed to sepolia. There is probably a more
hermetic solution, but this is the fastest way to get to the bottom of
this that I can figure out.

The token seems to be minted, but joining the space is failing, and it's
the IsSpaceMember chain auth kind that's failing. I'd like to get to the
bottom of why.