pull eventIds from snapshot when parsing sync reset (#2614)
these are used in the `initializeFromResponse` function in
syncedStream.ts
the result of this would be that usernames and dispaly names wouldn’t
decrypt imediately, they’d be pushed in the decryption queue. They items
would not be decrypted a second time, it would hit the cache again
before trying to decrypt, but still this clogs the queue and slows
everything down.