app: remove useLookupUsersWithAbstractAccountAddress (#9252)
this still causes the app to crash on Boku no Airdrop town even though
Austin put in a change for it a couple weeks ago.

I'm doing a sweep of our abstract account lookups, I'll have a bigger PR
later to get rid of it in most places, but this is probably the most
complex and unnecessary use of it in the app.

This hook was used for to append an abstract account to the user lookup
store data. This is very expensive if you have a space with a lot of
members, and there's no need to grab every single user's abstract
account address in a space.

This was used in 2 places: 
- the modal that pops up when you want to join a gated channel, but you
can't, and you see user addresses that are gating the town (Boku no
airdrop issue). We don't need to display AA addresses here
- in the gating UI for editing roles. We displayed the AA address next
to the username when you paste in a wallet address to gate by. I changed
this to remove the displayed AA address (which I plan on doing all over
the app), and added the avatar instead. This also is more in line with
the final list of users/wallets that is displayed before submitting the
form. The values that are submitted in the form don't change, they're
the same as before.

Removed AA address, added avatar
<img width="378" alt="Screenshot 2025-03-18 at 6 25 41 AM"
src="https://github.com/user-attachments/assets/44644a8f-b618-4c9a-b190-663f2f87ac22"
/>

Which matches this better
<img width="469" alt="Screenshot 2025-03-18 at 6 25 49 AM"
src="https://github.com/user-attachments/assets/834c6db8-8749-488d-8aaf-7fbd5e370701"
/>