E2E bot messaging in tests (#2603)
The events forwarding test now passes with the following sequence:

- a bot registers itself and webhook
- a chat participant creates a space and channel
- the bot adds itself to the channel
- the participants sends a message
- the app registry service detects that the message should be forwarded
to the bot, but that we don't have keys for the bot, and requests the
bot to solicit keys for the channel and session
- the test bot server makes the appropriate key solicitation
- the participant responds to the key solicitation by posting keys to
the bot's user inbox stream
- the app registry server detects the keys posted to the bot's user
inbox stream and forwards the message, along with a collection of
ciphertexts for the bot's device key that contains the session id the
message is encrypted under
- the bot service responds in channel with a reply that includes
ciphertext and session information to confirm that all needed metadata
was included in the forwarded message to the bot

This PR contains a bugfix in the StreamTrackerImpl where we were
previously using a context from a request to add a stream to the tracker
via AddStream, but this was an issue was the context was expiring after
the request ended, and thus the stream would no longer be tracked. The
tracker now stores it's creation context and uses that internally for
all AddStream calls.