node: move node to unified eventing model (#2649)
Switches the node over to use the new StreamUpdated event that was
introduced in https://github.com/towns-protocol/towns/pull/2578

The StreamUpdated event is a tuple of the type of event it represents
and abi.encoded data. The node decodes this data based on the event
type. For stream miniblock updates this is an slice of updates (batch)
but abi decoding in the node currently fails. A workaround is used that
json marshals the decoded input and marshals it back it which works.
This workaround is a todo that needs to be fixed.