app: trade analytics (towns-19083) (#9160)
not super fan of this* but I'm not sure there's an elegant way of doing
it without having to change the current structure.
Anyways, it's urgent so let's get it in a test it tomorrow, I've been
running though the scenarios I could think of already.

*) this = adding context via hooks where it's not necessarily needed if
not for the sake of analytics etc.

note: the fail event would probably not cover much at this point since
the errors are caught in the trading context - I'm going to re-iterate
here but in the meantime getting the difference between submit and
success should give an idea about the amount of failures.