Userops: modular account (#9311)
EDIT: There are 4,472 lines of new ABIs so under 2,000 lines of code to
actually look at 😅. And most of the code in
permissionless/accounts/modular doesn't need close review as it's ported
from other sdks.

This adds a [Modular
Account](https://github.com/alchemyplatform/modular-account) smart
account client implementation. The userop package is compatible with
userops from either SimpleAccount or ModularAccount.

Of particular note for other clients, @erikolsson @miguel-nascimento, is
the logic in
[determineSmartAccount](https://github.com/HereNotThere/harmony/blob/423e47ec51e465250a461949374ffcd161295092/clients/web/userops/src/lib/permissionless/accounts/determineSmartAccount.ts).
This logic is going to get ported to a stackup-worker api so that you
can call it in the iOS and staking site and receive the correct address
and other details for the smart account.

The idea is, you'd call this api, then set up the correct userop type
using that info - either a simple account or modular account compatible.

@erikolsson also note the `toModularAccount` and `nativeSMASIgner` files
- encoding calldata and signing is a bit more involved with the modular
account, and will have to be ported to iOS. We can go over this when
you're back/ready.

@miguel-nascimento basically you're gonna have to copy over all the code
in permissionless/accounts/modular, since permissionless does not
include this implementation out of the box.

This PR does not include the upgrade method, I'll add that later w/
tests.

The userop package reads from AA_NEW_ACCOUNT_IMPLEMENTATION_TYPE, which
in the app is hardcoded to "simple". Meaning nothing changes for our app
for now.

Tests:
All test work w/ both types of accounts, but in CI I'm only testing
simple account still b/c we're not upgrading yet.