replication: add support to StreamRegistry to migrate non-repl to repl streams (#2696)
This change allows gradual migration of existing non-replicated streams
to replicated streams. Migration is a 2 step process:
- first the stream nodes list is updated with the nodes the stream is
replicated on with a replication factor of 1. Only the first node on
this list will take part in quorum. The remaining nodes will only sync
the stream data into its local storage.
- bump the replication factor to the desired value. From this moment the
nodes will start participating in quorum and the stream is truly
replicated.