Update storage interface (#2705)
The goal of this PR is to prepare the storage interface for an upcoming
change: moving the snapshot out of the miniblock header and into a
separate DB column in both the miniblocks and miniblock_candidates
tables. Since this change will require updates across multiple
components and could take some time, it's a good idea to prepare as much
as possible now to minimize merge conflicts and make future changes
smoother.