userops: reduce CI time (#9330)
I found that you can initially start anvil with `--block-time 0.1` and
that works, but if you later want to update it with
`evm_setIntervalMining` as we do in some of our deploy scripts, that
only accepts an integer so you can never go back. So I extracted base
contract deployments (for now just duplicated) and removed some other
unnecessary stuff.

Used to take ~18mins, now 10