stackup-worker: add /api/smart-account and update wrangler, migrate to vitest (#9342)
This adds /api/smart-account which accepts
`newAccountImplementationType` and `ownerAddress` and returns `address`
which is the smart account address and accountType which is `simple` or
`modular`.

`newAccountImplementationType` should be either `simple` or `modular`,
and it should be whatever type new user accounts should get - so for now
this should be `simple`

I also went ahead and ported this to vitest, which required upgrading
wrangler, as Cloudflare no longer supports jest, and the test I was
porting from the client was in vitest already. It took longer than I
thought 😑 But at least we have an example for other workers.

Along w/ tests, I deployed this to alpha and tested the response in
web-app