Add ability to ban towns (#9338)
This PR adds support for banned towns. If a user tries to access one,
they’ll see a new BannedTownPage.tsx with a message letting them know
the town’s no longer available. There’s also a button to send them to
the explore page so they can check out other towns.

Handles all subpages for the town, remove the banned towns from sidebar
and explore, changes how it shows when town is shared in a different
town.

Also exposes the API so the env is available on iOS.