app: detect translation, fix fundwallet action button (#9369)
interesting bug i've not encountered before
https://github.com/facebook/react/issues/11538#issuecomment-350110297

User was trying to fund wallet, ErrorBoundary was triggering w/
`Failed to execute 'removeChild' on 'Node': The node to be removed is
not a child of this node.`

Reproducible if I translated site. I added some language logging in case
this causes issues elsewhere.

We could also disable translate for the site, but that's probably
overkill for now.