Add quotes around schema name for stats query function (#2737)
Uniquely generated random node ids for tests sometimes contain hyphens,
and these node ids are used to compose schema names for the database. If
the schema name contains a hyphen, then the schema must be quoted in
double quotes wherever it's referenced directly in queries.

This change updates the statistics calculation function to quote the
schema in double quotes everywhere it occurs.