Start counting tips sent and received “count” in user and member snapshots (#2735)
I initially added a tips sent/received amount, but we want the more
egalitarian count to take center stage. This pr zeros out the sent and
received amounts so that they will match… sadly this will now not match
the leaderboard, perhaps we can reset that too.

I really wish there was a good way to backfill this... but I can't think
of a good option at the moment.