Implement GetStatus webhook call with caching (#2731)
The GetStatus endpoint should return the encryption device of the bot
service along with the framework version, which is obtained by sending a
status request to the webhook registered for the bot. Additionally we
add a 2 second cache for status responses to protect the bot from DDOSes
via this endpoint.