Advance multi_ne RPC port (#2797)
Currently the justfile assigns a port to the app registry that is
RPC_PORT + NUM_INSTANCES, taking the next available port in sequence
after all stream nodes are assigned. This was causing an issue, however,
because the RPC_PORT env setting for multi and multi_ne are apart by
exactly 10, meaning that the app registry launched with the multi
configuration attempts to take the same port as node 0 of the multi_ne
config. There are many possible solutions here, but this one simply
advances the lowest assigned RPC_PORT for multi_ne config by 10 to
create more headroom.