node: fetch latest miniblock during reconcile for to-repl-streams migration (#2827)
For non-replicated streams nodes only register miniblocks periodically
to save on transaction costs. That means that the stream update event as
emitted during stream (re)placement contains a miniblock number that
might not be the latest and the node managing the stream at the moment
may have newer miniblocks.

In that case ask the existing quorum node for the latest block and
reconcile to that block.