Instrument the test app server to fail tests on certain errors (#2840)
The test bot server may encounter a variety of errors that indicate
either that the app registry service has a critical error in the
implementation, or that something failed during the test which is
causing the test not to cover what was expected. In these cases, we can
now send an error signal that will cause tests to fail.

Now that we fail tests on errors in the bot service, I noticed that we
were seeing membership events but not properly parsing them, and added
handling for that in the test bot server with a standard membership
event reply to show what events the bot was observing.