app: add VITE_BASE_RPC_GATEWAY_URL (#9473)
if you add VITE_BASE_RPC_GATEWAY_URL in render, it's going to use that
to construct the provider that is passed to the SpaceDapp.
if present, it will apply the VITE_BASE_RPC_GATEWAY_SAMPLING_RATE else a
0% sampling rate

These vars are used to compase the `gatewayBaseChain`. essentially, this
is equivalent to our existing `baseChain` passed to the
`TownsContextProvider`, except the rpcUrl is swapped out.

The existing `baseChain` is retained as it is applied to our account
abstaction config - meaning userop flows will not be using the gateway
url at all. Theoretically, the gateway rpc url is going to be used
strictly for reads from SpaceDapp.