Snapshot migration to flush lost session ids (#2925)
some usernames are just lost. users uploaded their username to the
stream but didn’t share any keys.
then everyone requests the same session id to decrypt, but nobody has
it, so we just save those requests forever.

this pr detects and flushes out lost username and displaynames.

We’re working on a real fix for this, but in the meantime this cuts the
ax1 snapshot size in half.

Snapshot size reduced from 42133234 to 16167148, 24.76 MB saved 409854
sessionIds 85 usernames and 0 display names removed