Fix potential erroneous false entitlement check result (#2909)
If a nested entitlement contains two on-chain checks and chain
unavailability causes rpc requests to lag until the context times out,
these could have erroneously come back as a false entitlement result
instead of an error result. I think it's very likely this would happen
any time a chain or rpc provider is partially unavailable. This PR
repairs the error for both AND and OR operators in entitlement trees.