chore: move `@connectrpc/connect-node` to peer dependency (#2938)
This allow us to correctly handle and support both web and node
environments.

Previously, this was making web clients downloading (and building!) node
dependencies.

Now, if you're building a node app, you should add
`@connectrpc/connect-node` as a dependency. It's not included on the sdk
anymore.



This will help us to move to the future of removing the esbuild from the
buildstep of our apps/bots.
If you're on web, you should add `@connectrpc/connect-web`.