web3: organize into contract directories (#2947)
This PR organizes the `@web3` package - it was pretty all over the
place.
It moves files into mostly contract related modules
```
├── airdrop/
├── base-registry/
├── cache/
├── delegate-registry/
├── eip-712/
├── erc-721/
├── platform-requirements/
├── pricing-modules/
├── river-registry/
├── space/
├── space-dapp/
├── space-owner/
├── space-registrar/
├── test-helpers/
├── types/
├── utils/
├── wallet-link/
├── BaseContractShim.ts
├── chain.ts
└── index.ts
```
- Core contract implementations (`space-owner/`, `space-registrar/`,
`wallet-link/`, etc.)
- `space/` contains most of the shims, as they correspond to facets on
the space
- `space-dapp/` is still the entrypoint for interacting with contracts
- Utility modules (`utils/`, `test-helpers/`, `cache/`, `types/`, etc.)
- Protocol-specific implementations (`eip-712/`, `erc-721/`)
- Base classes and shared functionality (`BaseContractShim.ts`,
`chain.ts`)

There is some redundancy and/or confusing logic to some of the shims.
That I will continue to address.

Notably, I swapped `ISpaceOwner` abi for `SpaceOwner` abi, including the
generated types in this PR. The `SpaceOwner.ts` class was using both
`IERC721AShim` and `ISpaceOwnerShim`, when in fact the abi for
`SpaceOwner` provides all relevant methods, and is the accurate
representation for interacting with the SpaceOwner contract.

Also, I moved `new SpaceOwner()` out of `Space.ts` and into
`SpaceDapp.ts`. `Space.ts` is a representation of an onchain space with
all of its facets - SpaceOwner is not a facet on spaces, but rather a
separate contract.

### Regarding Tests
We have a somewhat odd setup where you can't really write an integration
test in web3 w/out duplicating a bunch of code from sdk. You're somewhat
forced to write SpaceDapp tests within the sdk package. We should
probably move the required code (I believe mostly around
BaseChainConfig/RiverChainConfig) into another package that can be
shared between the two.