Proposal for modified handling of isNewDevice
I think it was a mistake to not pass session ids in the initial key solicitations, we end up in a situation where we can ping pong solicitations without actually satisfying the request. The contents of this change

- isNewDevice becomes a suggestion instead of a state
- all key requests now require session ids
- if isNewDevice is true, clients will still send all known sessions, but explicitly state in the fulfillment if they fulfilled any of the requested ids.

this should tighten the reqest loop and prevent at least one back and forth between clients in some cases.