Check client availability in remote syncer (#3021)
A remote node might be down so the remote syncer creation process is
going to be timed out in this case. As a result, the entire modify sync
request is going to fail.

This PR adds a logic to make sure the remote node is available by
sending the info request with a short timeout. Fixed a couple of bugs in
the remote syncer, namely keep alive mechanism and node unavailability
handling.